+++ internal/config/config.go
@@ -3,7 +3,7 @@ package config
 import (
 	"fmt"
 	"github.com/spf13/viper"
-	"github.com/sptuan/stargazer/modules/global"
+	"github.com/sptuan/stargazer/internal/global"
 	"path/filepath"
 	"strings"
 )
