+++ internal/global/global.go
@@ -1,6 +1,6 @@
 package global
 
-import "github.com/sptuan/stargazer/modules/model"
+import "github.com/sptuan/stargazer/internal/model"
 
 var ServerConfig model.ServerConfig
 var Targets []model.Target
