+++ internal/model/detector.go
