+++ internal/model/record.go
