+++ pkg/detector/http_detector.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"fmt"
 	"github.com/google/uuid"
-	"github.com/sptuan/stargazer/modules/logger"
-	"github.com/sptuan/stargazer/modules/model"
+	"github.com/sptuan/stargazer/internal/model"
+	"github.com/sptuan/stargazer/pkg/logger"
 	"io/ioutil"
 	"math"
 	"net/http"
