+++ pkg/detector/http_test.go
@@ -1,8 +1,8 @@
 package detector
 
 import (
-	"github.com/sptuan/stargazer/modules/logger"
-	"github.com/sptuan/stargazer/modules/model"
+	"github.com/sptuan/stargazer/internal/model"
+	"github.com/sptuan/stargazer/pkg/logger"
 	"testing"
 	"time"
 )
